app/requirements.txt
streamlit==1.31.1
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
plotly==5.18.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dateutil==2.8.2
pytz==2024.1
yfinance>=0.2.50
requests==2.31.0
python-dotenv==1.0.0
curl_cffi>=0.5.10